+++ soulng/slg/CppLiteralParser.cpp
@@ -21,7 +21,7 @@ soulng::parser::Match CppLiteralParser::CppLiteral(LexerFileLexer& lexer)
     }
     #endif // SOULNG_PARSER_DEBUG_SUPPORT
     soulng::parser::Match match(false);
-    int pos = lexer.GetPos();
+    int64_t pos = lexer.GetPos();
     soulng::lexer::Span span = lexer.GetSpan();
     switch (*lexer)
     {
